requirements.txt
arabic-reshaper==3.0.0
asgiref==3.9.1
asn1crypto==1.5.1
attrs==25.3.0
babel==2.17.0
certifi==2025.7.9
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.5
cssselect2==0.8.0
diff-match-patch==20241021
dj-database-url==3.0.1
Django==5.2.4
django-cors-headers==4.7.0
django-extensions==4.1
django-hordak==2.0.0
django-import-export==4.3.9
django-js-asset==3.1.2
django-money==3.5.4
django-mptt==0.18.0
django-treebeard==4.7.1
django-unfold==0.62.0
djangorestframework==3.16.0
drf-spectacular==0.28.0
et_xmlfile==2.0.0
Faker==37.5.3
html5lib==1.1
idna==3.10
inflection==0.5.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
lxml==6.0.0
Markdown==3.8.2
ofxtools==0.9.5
openpyxl==3.1.5
oscrypto==1.3.0
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.3.0
psycopg2-binary==2.9.10
py-moneyed==3.0
pycparser==2.22
pyHanko==0.29.1
pyhanko-certvalidator==0.27.0
pypdf==5.7.0
pypdfium2==4.30.0
python-bidi==0.6.6
python-dateutil==2.9.0.post0
PyYAML==6.0.2
referencing==0.36.2
reportlab==4.4.2
requests==2.32.4
rpds-py==0.27.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
svglib==1.5.1
tablib==3.8.0
tinycss2==1.4.0
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
uritools==5.0.0
urllib3==2.5.0
webencodings==0.5.1
xhtml2pdf==0.2.17
